requirements.txt
opencv-python
mediapipe
gradio
numpy
Pillow
torch
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git
```
